Deletar funcionando, e alterar.

+++ DevFlixN/src/main/java/br/upe/devflix/base/Entidade.java
@@ -14,32 +14,29 @@ public abstract class Entidade {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "id")
-	private int id;
+	private Long id;
 	private LocalDateTime dataInclusao;
 	private LocalDateTime dataUltimaAlteracao;
-
-	public int getId() {
+	
+	public Long getId() {
 		return id;
 	}
-
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
-
 	public LocalDateTime getDataInclusao() {
 		return dataInclusao;
 	}
-
 	public void setDataInclusao(LocalDateTime dataInclusao) {
 		this.dataInclusao = dataInclusao;
 	}
-
 	public LocalDateTime getDataUltimaAlteracao() {
 		return dataUltimaAlteracao;
 	}
-
 	public void setDataUltimaAlteracao(LocalDateTime dataUltimaAlteracao) {
 		this.dataUltimaAlteracao = dataUltimaAlteracao;
 	}
+
+	
 
 }
